+++ .Samples/MyApp/MyPreinstall.sh
@@ -1,2 +1,0 @@
-#!/bin/zsh
-echo "Hello from MyPreinstall.sh!"