+++ Samples/MyWallpaperConfigurator/PrepareDirectory.sh
@@ -1,0 +1,9 @@
+#!/bin/zsh
+
+if [ ! -d "/Library/Desktop Pictures/" ]; then
+    mkdir "/Library/Desktop Pictures/"
+fi
+
+if [ -f "/Library/Desktop Pictures/Snow Leopard Server.jpg" ]; then
+    rm "/Library/Desktop Pictures/Snow Leopard Server.jpg"
+fi